+++ elements/src/components/pos-switch/pos-switch.tsx
@@ -1,6 +1,7 @@
-import { Thing } from '@pod-os/core';
+import { RdfType, Relation, Thing } from '@pod-os/core';
 import { Component, Element, Event, h, Host, State } from '@stencil/core';
 import { ResourceAware, ResourceEventEmitter, subscribeResource } from '../events/ResourceAware';
+import { combineLatest, firstValueFrom, Observable, Subject, takeUntil } from 'rxjs';
 
 /**
  * Selects a child template to render based on properties of the subject resource, usually defined by an ancestor `pos-resource` element.
@@ -15,30 +16,36 @@ export class PosSwitch implements ResourceAware {
   @Element() host: HTMLElement;
   @State() error: string = null;
   @State() resource: Thing;
-  @State() caseElements: NodeListOf<HTMLPosCaseElement>;
-  @State() templateString: string;
+  @State() caseElements: HTMLPosCaseElement[];
+  @State() types: RdfType[];
+  @State() relations: Relation[];
+  @State() reverseRelations: Relation[];
+
+  private readonly disconnected$ = new Subject<void>();
 
   @Event({ eventName: 'pod-os:resource' })
   subscribeResource: ResourceEventEmitter;
 
   componentWillLoad() {
-    subscribeResource(this);
-
     const caseElements = this.host.querySelectorAll('pos-case');
     if (caseElements.length == 0) {
       this.error = 'No pos-case elements found';
     } else {
-      this.caseElements = caseElements;
+      this.caseElements = Array.from(caseElements);
+      subscribeResource(this);
     }
   }
 
   test(caseElement): boolean {
     let state = null;
     if (caseElement.getAttribute('if-typeof') !== null) {
-      state = this.resource
-        .types()
-        .map(x => x.uri)
-        .includes(caseElement.getAttribute('if-typeof'));
+      state = this.types.map(x => x.uri).includes(caseElement.getAttribute('if-typeof'));
+    }
+    if (caseElement.getAttribute('if-property') !== null) {
+      state = this.relations.map(x => x.predicate).includes(caseElement.getAttribute('if-property'));
+    }
+    if (caseElement.getAttribute('if-rev') !== null) {
+      state = this.reverseRelations.map(x => x.predicate).includes(caseElement.getAttribute('if-rev'));
     }
     if (caseElement.getAttribute('not') != null) {
       state = !state;
@@ -47,7 +54,32 @@ export class PosSwitch implements ResourceAware {
   }
 
   receiveResource = (resource: Thing) => {
-    this.resource = resource;
+    // reset any existing resource
+    this.disconnected$.next();
+    this.resource = undefined;
+    let observables: Observable<any>[] = [];
+    if (this.caseElements.some(caseElement => caseElement.hasAttribute('if-typeof'))) {
+      const observeTypes = resource.observeTypes().pipe(takeUntil(this.disconnected$));
+      observeTypes.subscribe(types => {
+        this.types = types;
+      });
+      observables.push(observeTypes);
+    }
+    if (this.caseElements.some(caseElement => caseElement.hasAttribute('if-property'))) {
+      const observeRelations = resource.observeRelations().pipe(takeUntil(this.disconnected$));
+      observeRelations.subscribe(relations => {
+        this.relations = relations;
+      });
+      observables.push(observeRelations);
+    }
+    if (this.caseElements.some(caseElement => caseElement.hasAttribute('if-rev'))) {
+      const observeReverseRelations = resource.observeReverseRelations().pipe(takeUntil(this.disconnected$));
+      observeReverseRelations.subscribe(reverseRelations => {
+        this.reverseRelations = reverseRelations;
+      });
+      observables.push(observeReverseRelations);
+    }
+    firstValueFrom(combineLatest(observables)).then(() => (this.resource = resource));
   };
 
   render() {
@@ -72,5 +104,10 @@ export class PosSwitch implements ResourceAware {
     });
     const activeElementsContent = activeElements.map(el => el.querySelector('template').innerHTML).join('\n');
     return <Host innerHTML={activeElementsContent}></Host>;
+  }
+
+  disconnectedCallback() {
+    this.disconnected$.next();
+    this.disconnected$.unsubscribe();
   }
 }